[Moved] the mobile register form to its own component
