+++ Client/src/assets/styling/components/_blueLine.scss
@@ -1,0 +1,6 @@
+.blue-line {
+    width: 100vw;
+    height: 2px;
+    background: linear-gradient(to right, #0d3ef100, #0D31F1, #0d3ef100);
+    margin: auto;
+}
