+++ Client/src/assets/styling/components/_consumerMain.scss
@@ -96,13 +96,6 @@ main {
         }
     }
 
-    .blue-line {
-        margin-top: 60px;
-        width: 100vw;
-        height: 2px;
-        background: linear-gradient(to right, #0d3ef100, #0D31F1, #0d3ef100);
-    }
-
     .stressed-man-img {
         margin-top: 30px;
         max-width: 100vw;
@@ -245,9 +238,6 @@ main {
             }
         }
         
-        .blue-line {
-            margin-top: 30px;
-        }
     }
 }
 
@@ -318,8 +308,5 @@ main {
             }
         }
         
-        .blue-line {
-            margin-top: 40px;
-        }
     }
 }
