+++ Client/src/assets/styling/components/_consumerNav.scss
@@ -61,7 +61,7 @@ nav {
 
                 &:hover {
                     transition: all 0.2s ease-in-out;
-                    background-color: $el-active-blue;
+                    border: 2px solid $el-active-blue;
                 }
             }
