--- Client/src/assets/styling/components/_registerForm.scss
@@ -1,146 +1,0 @@
-@use '../abstracts/colors' as *;
-@use '../abstracts/fonts' as *;
-
-.register-nav {
-    display: flex;
-    align-items: center;
-    gap: 20px;
-    padding: 0;
-    padding: 16px;
-
-    .router-link {
-        color: $text-main;
-        width: 45px;
-        height: 45px;
-        border-radius: 12px;
-        border: 2px solid $el-border-gray;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-    }
-
-    .btn-back {
-        transition: all 0.2s ease-in-out;
-        background: transparent;
-        color: $text-main;
-        width: 45px;
-        height: 45px;
-        border-radius: 12px;
-        border: 2px solid $el-border-gray;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-
-        &:hover {
-            transition: all 0.2s ease-in-out;
-            border: 2px solid $el-active-blue;
-            color: $el-active-blue;
-            cursor: pointer;
-        }
-    }
-
-    h1 {
-        color: $text-main;
-        font-family: map-get($font-title-black, font-family);
-        font-weight: map-get($font-title-black, font-weight);
-    }
-}
-
-form {
-    padding: 16px;
-    margin-top: 80px;
-    display: flex;
-    flex-direction: column;
-
-    label {
-        color: $text-main;
-        font-family: map-get($font-text-light, font-family);
-        font-weight: map-get($font-text-light, font-weight);
-    }
-
-    input {
-        position: relative;
-        transition: all 0.2s ease-in-out;
-        margin: 5px 0 10px 0;
-        width: calc(100vw - 32px);
-        height: 42px;
-        background-color: $el-bg-gray;
-        color: $text-main;
-        border: 2px solid $el-border-gray;
-        padding: 10px;
-        border-radius: 10px;
-        z-index: 1;
-
-        &:focus {
-            transition: all 0.2s ease-in-out;
-            outline: none;
-            border: 2px solid transparent;
-            border-radius: 10px;
-            background-image: linear-gradient($el-border-gray-focus, $el-border-gray-focus), 
-            linear-gradient(
-                to bottom,
-                rgb(13, 63, 241),
-                rgb(13, 63, 241, 0.2)
-              );
-            background-origin: border-box;
-            background-clip: padding-box, border-box;
-        }
-    }
-
-    .input-error {
-        transition: all 0.2s ease-in-out;
-        outline: none;
-        border: 2px solid transparent;
-        border-radius: 10px;
-        background-image: linear-gradient($el-border-gray-focus, $el-border-gray-focus), 
-        linear-gradient(
-            to bottom,
-            rgb(241, 13, 13),
-            rgb(241, 13, 13, 0.2)
-          );
-        background-origin: border-box;
-        background-clip: padding-box, border-box;
-
-        &:focus {
-            background-image: linear-gradient($el-border-gray-focus, $el-border-gray-focus), 
-            linear-gradient(
-                to bottom,
-                rgb(241, 13, 13),
-                rgb(241, 13, 13, 0.2)
-              );
-        }
-    }
-
-    p {
-        color: $el-error-red;
-        margin-bottom: 20px;
-    }
-
-    button {
-        transition: all 0.2s ease-in-out;
-        background-color: $el-active-blue;
-        border: none;
-        color: $text-main;
-        font-family: map-get($font-title-bold, font-family);
-        font-weight: map-get($font-title-bold, font-weight);
-        height: 35px;
-        border-radius: 50px;
-        margin-top: 10px;
-
-        &:hover {
-            transition: all 0.2s ease-in-out;
-            cursor: pointer;
-            background-color: $el-active-blue-hover;
-        }
-    }
-}
-
-.text-form-container {
-    padding: 60px 16px;
-
-    .router-link-text, p {
-        color: $text-main;
-        font-family: map-get($font-text-light, font-family);
-        font-weight: map-get($font-text-light, font-weight);
-    }
-}
