+++ Client/src/assets/styling/components/_registerFormMobile.scss
@@ -1,0 +1,134 @@
+@use '../abstracts/colors' as *;
+@use '../abstracts/fonts' as *;
+
+.form-background {
+    padding: 16px;
+
+    .mobile-form-nav {
+        display: flex;
+        align-items: center;
+        gap: 20px;
+        padding: 0;
+
+        .router-link {
+            color: $text-main;
+            width: 40px;
+            height: 40px;
+            border-radius: 12px;
+            border: 2px solid $el-border-gray;
+            display: flex;
+            align-items: center;
+            justify-content: center;
+        }
+
+        h2 {
+            font-size: 1.5rem;
+            color: $text-main;
+            font-family: map-get($font-title-black, font-family);
+            font-weight: map-get($font-title-black, font-weight);
+        }
+    }
+
+    .mobile-register-form {
+        margin-top: 50px;
+        display: flex;
+        flex-direction: column;
+
+        label {
+            color: $text-main;
+            font-family: map-get($font-text-light, font-family);
+            font-weight: map-get($font-text-light, font-weight);
+        }
+
+        input {
+            position: relative;
+            transition: all 0.2s ease-in-out;
+            margin: 10px 0 20px 0;
+            width: calc(100vw - 32px);
+            height: 42px;
+            background-color: $el-bg-gray;
+            color: $text-main;
+            border: 2px solid $el-border-gray;
+            padding: 10px;
+            border-radius: 10px;
+            z-index: 1;
+    
+            &:focus {
+                transition: all 0.2s ease-in-out;
+                outline: none;
+                border: 2px solid transparent;
+                border-radius: 10px;
+                background-image: linear-gradient($el-border-gray-focus, $el-border-gray-focus), 
+                linear-gradient(
+                    to bottom,
+                    rgb(13, 63, 241),
+                    rgb(13, 63, 241, 0.2)
+                  );
+                background-origin: border-box;
+                background-clip: padding-box, border-box;
+            }
+        }
+
+        .input-error {
+            transition: all 0.2s ease-in-out;
+            outline: none;
+            border: 2px solid transparent;
+            border-radius: 10px;
+            background-image: linear-gradient($el-border-gray-focus, $el-border-gray-focus), 
+            linear-gradient(
+                to bottom,
+                rgb(241, 13, 13),
+                rgb(241, 13, 13, 0.2)
+              );
+            background-origin: border-box;
+            background-clip: padding-box, border-box;
+    
+            &:focus {
+                background-image: linear-gradient($el-border-gray-focus, $el-border-gray-focus), 
+                linear-gradient(
+                    to bottom,
+                    rgb(241, 13, 13),
+                    rgb(241, 13, 13, 0.2)
+                  );
+            }
+        }
+
+        p {
+            color: $el-error-red;
+            margin-bottom: 20px;
+            margin-top: -10px;
+        }
+    
+        .register-mobile-button {
+            transition: all 0.2s ease-in-out;
+            background-color: $el-active-blue;
+            border: none;
+            color: $text-main;
+            font-family: map-get($font-title-bold, font-family);
+            font-weight: map-get($font-title-bold, font-weight);
+            height: 35px;
+            border-radius: 50px;
+            margin: 10px 0 32px 0;
+    
+            &:hover {
+                transition: all 0.2s ease-in-out;
+                cursor: pointer;
+                background-color: $el-active-blue-hover;
+            }
+        }
+    }
+
+    .text-form-container {
+        padding: 32px 0;
+    
+        .router-link-text, p {
+            color: $text-main;
+            font-family: map-get($font-text-light, font-family);
+            font-weight: map-get($font-text-light, font-weight);
+        }
+
+        p {
+            margin-bottom: 10px;
+        }
+    }
+}
