+++ Client/src/assets/styling/main.scss
@@ -4,6 +4,7 @@
 @import './components/consumerHeader';
 @import './components/consumerMain';
 @import './components/consumerFooter';
-@import './components/registerForm';
+@import './components/registerFormMobile';
 @import './components/registerDialog';
-@import './components/dialog'
+@import './components/dialog';
+@import './components/blueLine'
